Move exception assertion tests to a dedicated test class.
