+++ src/Fixie.Tests/Assertions/AssertionTests.cs
@@ -234,88 +234,6 @@ public class AssertionTests
         Contradiction(a1, value => value.Should(x => x == a2), "value should be == a2 but was Fixie.Tests.Assertions.AssertionTests+SampleA");
     }
 
-    public void ShouldAssertExpectedExceptions()
-    {
-        var doNothing = () => { };
-        Action divideByZero = () => throw new DivideByZeroException("Divided By Zero");
-
-        divideByZero.ShouldThrow<DivideByZeroException>("Divided By Zero")
-            .ShouldBe<DivideByZeroException>();
-
-        divideByZero.ShouldThrow<Exception>("Divided By Zero")
-            .ShouldBe<DivideByZeroException>();
-
-        Contradiction(doNothing, noop => noop.ShouldThrow<DivideByZeroException>("Divided By Zero"),
-            """
-            noop should have thrown System.DivideByZeroException but did not
-            """);
-
-        Contradiction(divideByZero, divide => divide.ShouldThrow<DivideByZeroException>("Divided By One Minus One"),
-            """
-            divide should have thrown System.DivideByZeroException with message
-            
-                "Divided By One Minus One"
-            
-            but instead the message was
-            
-                "Divided By Zero"
-            """);
-
-        Contradiction(divideByZero, divide => divide.ShouldThrow<ArgumentNullException>("Argument Null"),
-            """
-            divide should have thrown System.ArgumentNullException with message
-            
-                "Argument Null"
-            
-            but instead it threw System.DivideByZeroException with message
-            
-                "Divided By Zero"
-            """);
-    }
-
-    public async Task ShouldAssertExpectedAsyncExceptions()
-    {
-        var doNothing = async () => { await Task.CompletedTask; };
-        var divideByZero = async () =>
-        {
-            await Task.CompletedTask;
-            throw new DivideByZeroException("Divided By Zero");
-        };
-
-        var actualDivideByZeroException = await divideByZero.ShouldThrow<DivideByZeroException>("Divided By Zero");
-        actualDivideByZeroException.ShouldBe<DivideByZeroException>();
-
-        var actualException = await divideByZero.ShouldThrow<Exception>("Divided By Zero");
-        actualException.ShouldBe<DivideByZeroException>();
-
-        await Contradiction(doNothing, noop => noop.ShouldThrow<DivideByZeroException>("Divided By Zero"),
-            """
-            noop should have thrown System.DivideByZeroException but did not
-            """);
-
-        await Contradiction(divideByZero, divide => divide.ShouldThrow<DivideByZeroException>("Divided By One Minus One"),
-            """
-            divide should have thrown System.DivideByZeroException with message
-            
-                "Divided By One Minus One"
-
-            but instead the message was
-            
-                "Divided By Zero"
-            """);
-
-        await Contradiction(divideByZero, divide => divide.ShouldThrow<ArgumentNullException>("Argument Null"),
-            """
-            divide should have thrown System.ArgumentNullException with message
-            
-                "Argument Null"
-
-            but instead it threw System.DivideByZeroException with message
-            
-                "Divided By Zero"
-            """);
-    }
-
     class SampleA;
     class SampleB;
